requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.8.3
click==8.2.1
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.12.0
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
aiosqlite==0.20.0
greenlet==2.0.4
rich==13.7.1
